+++ src/test/java/zmq/TermEndpointIpcTest.java
@@ -1,6 +1,8 @@
 package zmq;
 
-import java.util.UUID;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
 
 public class TermEndpointIpcTest extends TestTermEndpoint
 {
@@ -11,8 +13,9 @@ public class TermEndpointIpcTest extends TestTermEndpoint
     }
 
     @Override
-    protected String endpointNormal()
-    {
-        return "ipc://" + UUID.randomUUID();
+    protected String endpointNormal() throws IOException {
+        Path temp = Files.createTempFile("zmq-test-", ".sock");
+        Files.delete(temp);
+        return "ipc://" + temp;
     }
 }
